Tools/Spreadsheet_Maker/requirements.txt
pandas>=2.0.3
openpyxl>=3.1.2
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0 